+++ src/graphql/entity/create_field/resolve_nested/get_parent_value/mod.rs
@@ -1,5 +1,5 @@
 use async_graphql::{ErrorExtensions, Value};
-use log::debug;
+use log::{debug, error, trace};
 
 use crate::graphql::entity::ServiceEntity;
 
@@ -16,12 +16,13 @@ impl ServiceEntity {
         let parent_value = match parent_value.clone().into_json() {
             Ok(value) => value,
             Err(_) => {
+                error!("Invalid JSON Object - Failed to convert to JSON");
                 return Err(async_graphql::Error::new(
                     "Invalid JSON Object - Failed to convert to JSON",
                 )
                 .extend_with(|_err, e| {
                     e.set("field", field_name);
-                }))
+                }));
             }
         };
 
@@ -29,31 +30,33 @@ impl ServiceEntity {
             let json_str = match parent_value.as_str() {
                 Some(value) => value,
                 None => {
+                    error!("Invalid JSON Object - Failed to convert to JSON string");
                     return Err(async_graphql::Error::new(
                         "Invalid JSON Object - Failed to convert to JSON string",
                     )
                     .extend_with(|_err, e| {
                         e.set("field", field_name);
-                    }))
+                    }));
                 }
             };
 
             match serde_json::from_str(&json_str) {
                 Ok(value) => value,
                 Err(_) => {
+                    error!("Invalid JSON Object - Failed to convert to JSON object");
                     return Err(async_graphql::Error::new(
                         "Invalid JSON Object - Failed to convert to JSON object",
                     )
                     .extend_with(|_err, e| {
                         e.set("field", field_name);
-                    }))
+                    }));
                 }
             }
         } else {
             parent_value
         };
 
-        debug!("Parent Value: {:?}", parent_value);
+        trace!("Parent Value: {:?}", parent_value);
 
         Ok(parent_value)
     }