+++ src/graphql/entity/create_field/resolve_root/mod.rs
@@ -1,7 +1,7 @@
 use async_graphql::{dynamic::ResolverContext, ErrorExtensions, Value};
 use bson::Document;
 use json::JsonValue;
-use log::{debug, error};
+use log::{debug, error, trace};
 
 use crate::{
     configuration::subgraph::entities::service_entity_field::ServiceEntityFieldConfig,
@@ -17,32 +17,39 @@ impl ServiceEntity {
         entity_field: &ServiceEntityFieldConfig,
         entity_required: bool,
     ) -> Result<Option<Value>, async_graphql::Error> {
+        debug!("Resolve Root Field");
         let field_name = ctx.field().name();
-
-        debug!("Resolving Root Field: {:?}", &field_name);
+        trace!("Field Name: {:?}", field_name);
 
         let value = match data_source {
             DataSource::Mongo(_ds) => {
                 let doc = match ctx.parent_value.try_downcast_ref::<Option<Document>>() {
                     Ok(doc) => {
+                        trace!("Parent Value: {:?}", doc);
                         if let Some(doc) = doc {
+                            trace!("Document Matched: {:?}", doc);
                             let value = entity_field
                                 .scalar
                                 .clone()
                                 .document_field_to_async_graphql_value(doc, entity_field)?;
-                            Ok(Some(value))
+                            trace!("Resolved Value: {:?}", value);
+                            Ok(value)
                         } else {
                             if entity_required {
+                                error!(
+                                    "Failed to resolve root field. Entity is marked as required. This may be due to a join on a required entity."
+                                );
                                 return Err(async_graphql::Error::new(
-                                    "Failed to resolve root field.",
+                                    "Failed to resolve root field. Entity is marked as required. This may be due to a join on a required entity.",
                                 )
                                 .extend_with(|_err, e| {
                                     e.set("field", field_name);
                                     e.set("entity", entity_field.name.clone());
                                 }));
                             } else {
-                                debug!("Parent value is null, returning null.");
+                                trace!("Parent value is null, returning null.");
                                 if entity_required {
+                                    error!("Failed to resolve root field.");
                                     return Err(async_graphql::Error::new(
                                         "Failed to resolve root field.",
                                     )
@@ -51,7 +58,7 @@ impl ServiceEntity {
                                         e.set("entity", entity_field.name.clone());
                                     }));
                                 } else {
-                                    debug!("Parent value is null, returning null.");
+                                    trace!("Parent value is null, returning null.");
                                     Ok(Some(Value::Null))
                                 }
                             }
@@ -59,13 +66,14 @@ impl ServiceEntity {
                     }
                     Err(_) => {
                         if entity_required {
-                            return Err(async_graphql::Error::new("Failed to resolve root field.")
+                            error!("Failed to resolve root field. Entity is marked as required. This may be due to a join on a required entity.");
+                            return Err(async_graphql::Error::new("Failed to resolve root field. Entity is marked as required. This may be due to a join on a required entity.")
                                 .extend_with(|_err, e| {
                                     e.set("field", field_name);
                                     e.set("entity", entity_field.name.clone());
                                 }));
                         } else {
-                            debug!("Failed to downcast parent value, returning null.");
+                            trace!("Failed to downcast parent value, returning null.");
                             return Ok(Some(Value::Null));
                         }
                     }
@@ -78,7 +86,8 @@ impl ServiceEntity {
                     Ok(json_value) => json_value,
                     Err(_) => {
                         if entity_required {
-                            return Err(async_graphql::Error::new("Failed to resolve root field.")
+                            error!("Failed to resolve root field. Entity is marked as required. This may be due to a join on a required entity.");
+                            return Err(async_graphql::Error::new("Failed to resolve root field. Entity is marked as required. This may be due to a join on a required entity.")
                                 .extend_with(|_err, e| {
                                     e.set("field", field_name);
                                     e.set("entity", entity_field.name.clone());
@@ -105,29 +114,31 @@ impl ServiceEntity {
                             Ok(Some(value))
                         } else {
                             if entity_required {
-                                error!("Failed to resolve root field.");
+                                error!(
+                                    "Failed to resolve root field. Entity is marked as required. This may be due to a join on a required entity"
+                                );
                                 return Err(async_graphql::Error::new(
-                                    "Failed to resolve root field.",
+                                    "Failed to resolve root field. Entity is marked as required. This may be due to a join on a required entity.",
                                 )
                                 .extend_with(|_err, e| {
                                     e.set("field", field_name);
                                     e.set("entity", entity_field.name.clone());
                                 }));
                             } else {
-                                debug!("Parent value is null, returning null.");
+                                trace!("Parent value is null, returning null.");
                                 Ok(Some(Value::Null))
                             }
                         }
                     }
                     Err(e) => {
-                        debug!("Failed to downcast parent value: {:?}", e);
+                        trace!("Failed to downcast parent value: {:?}", e);
                         if entity_required {
                             error!(
-                                "Failed to resolve root field, `{}`. Entity is marked as required.",
+                                "Failed to resolve root field, `{}`. Entity is marked as required. This may be due to a join on a required entity.",
                                 entity_field.name
                             );
                             return Err(async_graphql::Error::new(format!(
-                                "Failed to resolve root field, `{}`. The entity is marked as required.",
+                                "Failed to resolve root field, `{}`. The entity is marked as required. This may be due to a join on a required entity.",
                                 entity_field.name
                             ))
                             .extend_with(|_err, e| {
@@ -143,7 +154,7 @@ impl ServiceEntity {
             }
         };
 
-        debug!("Resolved Root Field: {:?}", value);
+        trace!("Resolved Root Field: {:?}", value);
 
         value
     }