+++ src/graphql/entity/get_field_type_ref/create_internal_type_field_refs/mod.rs
@@ -34,6 +34,7 @@ impl ServiceEntity {
         TypeRefsAndDefs {
             type_ref,
             type_defs: Vec::new(),
+            enum_defs: Vec::new(),
         }
     }
 }